requirements.txt
Flask==2.3.2
tensorflow==2.13.0
gTTS==2.3.2
SpeechRecognition==3.8.1
playsound==1.2.2
requests==2.31.0
numpy==1.25.2
pillow==9.5.0